+++ src/model/packet.rs
@@ -40,7 +40,7 @@ use super::timeout::Timeout;
 /// **mac**
 /// 消息验证码。如果已经协商了消息验证，该域包含 MAC。初始时，MAC 算法必须是"none"。
 
-fn read_with_timeout<S>(stream: &mut S, tm: u64, buf: &mut [u8]) -> SshResult<()>
+fn read_with_timeout<S>(stream: &mut S, tm: u128, buf: &mut [u8]) -> SshResult<()>
 where
     S: Read,
 {
@@ -71,7 +71,7 @@ where
     }
 }
 
-fn try_read<S>(stream: &mut S, _tm: u64, buf: &mut [u8]) -> SshResult<usize>
+fn try_read<S>(stream: &mut S, _tm: u128, buf: &mut [u8]) -> SshResult<usize>
 where
     S: Read,
 {
@@ -87,7 +87,7 @@ where
     }
 }
 
-fn write_with_timeout<S>(stream: &mut S, tm: u64, buf: &[u8]) -> SshResult<()>
+fn write_with_timeout<S>(stream: &mut S, tm: u128, buf: &[u8]) -> SshResult<()>
 where
     S: Write,
 {
